move section style into components that use it
